requirements.txt
yolov5~=5.0.0
opencv-python~=4.5.3.56
Pillow~=8.3.2
numpy~=1.21.2
boto3~=1.18.40
dlib~=19.22.1